requirements.txt
# Core Cloud Function Framework
functions-framework==3.*

# Telegram Bot API
python-telegram-bot==20.*

# Google AI (Gemini Vision)
google-generativeai>=0.5.0

# Google Sheets API
gspread>=5.0
google-auth>=2.0

# HTTP Client
aiohttp>=3.8

# Fuzzy Matching
thefuzz>=0.22
python-Levenshtein>=0.25

# Timezone handling
pytz>=2024.1

# Testing
pytest>=8.0
pytest-asyncio>=0.23
